+++ src/Lux.jl
@@ -46,6 +46,8 @@ include("autodiff.jl")
 function __init__()
     @require Flux="587475ba-b771-5e3f-ad9e-33799f191a9c" include("transform.jl")
 end
+# Contrib
+include("contrib/emptyarrays.jl")
 
 # Deprecations
 include("deprecated.jl")